공통 hook으로 변경
