+++ src/renderer/shared/hooks/index.ts
@@ -5,3 +5,4 @@ export * from './use-auth-token';
 export * from './use-auth-client';
 export * from './use-disclosure';
 export * from './use-notification';
+export * from './use-rive-cat';
