+++ src/renderer/widgets/pomodoro/ui/home-screen.tsx
@@ -1,10 +1,8 @@
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 
-import { StateMachineInput, StateMachineInputType, useRive } from '@rive-app/react-canvas';
 import { useNavigate } from 'react-router-dom';
 import { useLocalStorage } from 'usehooks-ts';
 
-import { CatType } from '@/entities/cat';
 import { PomodoroMode } from '@/entities/pomodoro';
 import { useCategories, useUpdateCategory, ChangeCategoryDrawer } from '@/features/category';
 import { catNameMap } from '@/features/pomodoro';
@@ -12,7 +10,7 @@ import { ChangeTimeDialog } from '@/features/time';
 import { useUser } from '@/features/user';
 import catHomeMotionRiveFile from '@/shared/assets/rivs/cat_home.riv';
 import { LOCAL_STORAGE_KEY, PATH } from '@/shared/constants';
-import { useDisclosure } from '@/shared/hooks';
+import { useDisclosure, useRiveCat } from '@/shared/hooks';
 import { Button, Guide, Icon, Tooltip } from '@/shared/ui';
 import { getCategoryIconName, createIsoDuration } from '@/shared/utils';
 
@@ -20,12 +18,6 @@ const steps = [
   { id: 'categoryButton', message: '눌러서 카테고리를 변경할 수 있어요' },
   { id: 'timeAdjustDiv', message: '눌러서 시간을 조정할 수 있어요' },
 ];
-const RIVE_STATE_MACHINE_NAME = 'State Machine_Home';
-const userCatTypeAliasMap: Record<CatType, string> = {
-  CHEESE: 'cheese',
-  BLACK: 'black',
-  THREE_COLOR: 'calico',
-};
 
 type HomeScreenProps = {
   setMode: (mode: PomodoroMode) => void;
@@ -57,34 +49,11 @@ export const HomeScreen = ({
   const { mutate: updateCategory } = useUpdateCategory();
   const { data: user } = useUser();
 
-  const { rive, RiveComponent } = useRive({
+  const { RiveComponent, clickCatInput } = useRiveCat({
     src: catHomeMotionRiveFile,
-    stateMachines: RIVE_STATE_MACHINE_NAME,
+    stateMachines: 'State Machine_Home',
+    userCatType: user?.cat?.type,
   });
-  const [clickInput, setClickInput] = useState<StateMachineInput>();
-
-  useEffect(() => {
-    const userCatType = user?.cat?.type;
-    if (!rive || !userCatType) return;
-
-    const userCatTypeAlias = userCatTypeAliasMap[userCatType];
-    const inputs = rive.stateMachineInputs(RIVE_STATE_MACHINE_NAME);
-
-    const booleanInputs = inputs.filter((input) => input.type === StateMachineInputType.Boolean);
-    const triggerInputs = inputs.filter((input) => input.type === StateMachineInputType.Trigger);
-
-    // boolean input들 중 user의 고양이 타입에 해당하는 input만 true로 설정
-    booleanInputs.forEach((input) => {
-      input.value = input.name.toLowerCase().includes(userCatTypeAlias);
-    });
-    // trigger input들 중 user의 고양이 타입에 해당하는 input을 클릭할 수 있도록 설정
-    const clickInput = triggerInputs.find((input) => {
-      return input.name.toLowerCase().includes(userCatTypeAlias);
-    });
-    setClickInput(clickInput);
-
-    rive.play();
-  }, [rive, user?.cat?.type]);
 
   return (
     <>
@@ -109,7 +78,7 @@ export const HomeScreen = ({
           <RiveComponent
             className="w-full h-[240px]"
             onClick={() => {
-              clickInput?.fire();
+              clickCatInput?.fire();
             }}
           />
